+++ src/Mod/TechDraw/Gui/Resources/TechDraw.qrc
@@ -85,30 +85,43 @@
         <file>icons/rectangle.svg</file>
         <file>icons/square.svg</file>
         <file>icons/TechDraw_3PtAngleDimension.svg</file>
+        <file>icons/TechDraw_3PtAngleDimension_Pointer.svg</file>
         <file>icons/TechDraw_AngleDimension.svg</file>
+        <file>icons/TechDraw_AngleDimension_Pointer.svg</file>
         <file>icons/TechDraw_AreaDimension.svg</file>
+        <file>icons/TechDraw_AreaDimension_Pointer.svg</file>
         <file>icons/TechDraw_Balloon.svg</file>
         <file>icons/TechDraw_CameraOrientation.svg</file>
         <file>icons/TechDraw_DiameterDimension.svg</file>
+        <file>icons/TechDraw_DiameterDimension_Pointer.svg</file>
         <file>icons/TechDraw_Dimension.svg</file>
         <file>icons/TechDraw_Dimension_Pointer.svg</file>
         <file>icons/TechDraw_DimensionRepair.svg</file>
         <file>icons/TechDraw_ExtensionAreaAnnotation.svg</file>
         <file>icons/TechDraw_ExtensionArcLengthAnnotation.svg</file>
+        <file>icons/TechDraw_LengthArcDimension_Pointer.svg</file>
         <file>icons/TechDraw_ExtensionCascadeHorizDimension.svg</file>
         <file>icons/TechDraw_ExtensionCascadeObliqueDimension.svg</file>
         <file>icons/TechDraw_ExtensionCascadeVertDimension.svg</file>
         <file>icons/TechDraw_ExtensionChangeLineAttributes.svg</file>
         <file>icons/TechDraw_ExtensionCircleCenterLines.svg</file>
         <file>icons/TechDraw_ExtensionCreateHorizChainDimension.svg</file>
+        <file>icons/TechDraw_HorizontalChainDimension_Pointer.svg</file>
         <file>icons/TechDraw_ExtensionCreateHorizChamferDimension.svg</file>
+        <file>icons/TechDraw_HorizontalChamferDimension_Pointer.svg</file>
         <file>icons/TechDraw_ExtensionCreateHorizCoordDimension.svg</file>
+        <file>icons/TechDraw_HorizontalCoordDimension_Pointer.svg</file>
         <file>icons/TechDraw_ExtensionCreateLengthArc.svg</file>
         <file>icons/TechDraw_ExtensionCreateObliqueChainDimension.svg</file>
+        <file>icons/TechDraw_ObliqueChainDimension_Pointer.svg</file>
         <file>icons/TechDraw_ExtensionCreateObliqueCoordDimension.svg</file>
+        <file>icons/TechDraw_ObliqueCoordDimension_Pointer.svg</file>
         <file>icons/TechDraw_ExtensionCreateVertChainDimension.svg</file>
+        <file>icons/TechDraw_VerticalChainDimension_Pointer.svg</file>
         <file>icons/TechDraw_ExtensionCreateVertChamferDimension.svg</file>
+        <file>icons/TechDraw_VerticalChamferDimension_Pointer.svg</file>
         <file>icons/TechDraw_ExtensionCreateVertCoordDimension.svg</file>
+        <file>icons/TechDraw_VerticalCoordDimension_Pointer.svg</file>
         <file>icons/TechDraw_ExtensionCustomizeFormat.svg</file>
         <file>icons/TechDraw_ExtensionDecreaseDecimal.svg</file>
         <file>icons/TechDraw_ExtensionDrawCosmArc.svg</file>
@@ -136,9 +149,12 @@
         <file>icons/TechDraw_ExtensionThreadHoleSide.svg</file>
         <file>icons/TechDraw_ExtensionVertexAtIntersection.svg</file>
         <file>icons/TechDraw_HorizontalDimension.svg</file>
+        <file>icons/TechDraw_HorizontalDimension_Pointer.svg</file>
         <file>icons/TechDraw_HorizontalExtentDimension.svg</file>
+        <file>icons/TechDraw_HorizontalExtentDimension_Pointer.svg</file>
         <file>icons/TechDraw_LandmarkDimension.svg</file>
         <file>icons/TechDraw_LengthDimension.svg</file>
+        <file>icons/TechDraw_LengthDimension_Pointer.svg</file>
         <file>icons/TechDraw_LinkDimension.svg</file>
         <file>icons/TechDraw_Lock.svg</file>
         <file>icons/TechDraw_Pages.svg</file>
@@ -153,6 +169,7 @@
         <file>icons/TechDraw_ProjRight.svg</file>
         <file>icons/TechDraw_ProjTop.svg</file>
         <file>icons/TechDraw_RadiusDimension.svg</file>
+        <file>icons/TechDraw_RadiusDimension_Pointer.svg</file>
         <file>icons/TechDraw_TreeHatch.svg</file>
         <file>icons/TechDraw_TreeMulti.svg</file>
         <file>icons/TechDraw_TreePage.svg</file>
@@ -165,7 +182,9 @@
         <file>icons/TechDraw_TreeSymbol.svg</file>
         <file>icons/TechDraw_TreeView.svg</file>
         <file>icons/TechDraw_VerticalDimension.svg</file>
+        <file>icons/TechDraw_VerticalDimension_Pointer.svg</file>
         <file>icons/TechDraw_VerticalExtentDimension.svg</file>
+        <file>icons/TechDraw_VerticalExtentDimension_Pointer.svg</file>
         <file>icons/TechDraw_RefError.svg</file>
         <file>icons/triangle.svg</file>
